Change camelize from class method to lambda
I slightly prefer this, since Camelization is not really a concern of
the class's and thus does not really make sense as a class method.

In the future, I would like to make the mapping of names a pluggable
strategy, but I do not have that need right now. YAGNI and all that...